+++ onlinepayments-sdk/src/test/kotlin/com/onlinepayments/sdk/client/android/domain/paymentRequest/PaymentRequestFieldTest.kt
@@ -134,6 +134,6 @@ class PaymentRequestFieldTest {
             paymentRequestField.setValue("4222422242224222")
         }
 
-        assertEquals("Cannot write \"READ_ONLY\" field: cardNumber", exception.message)
+        assertEquals("Cannot write READ_ONLY field: cardNumber", exception.message)
     }
 }